refactor: 예외처리 코드 이동 및 수정
